price-extractor-python/requirements-test.txt
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-asyncio==0.21.1
pytest-timeout==2.2.0
pytest-xdist==3.3.1
pytest-env==1.0.1
requests-mock==1.11.0
beautifulsoup4==4.12.2
playwright==1.39.0
anthropic>=0.5.0
httpx==0.25.1
pytest-recording==0.13.0
vcrpy==5.1.0
freezegun==1.2.2 